fix: move execute button to actions section
