+++ Src/WitsmlExplorer.Frontend/components/ContentViews/QueryView/components/QueryOptions/QueryOptions.tsx
@@ -1,7 +1,12 @@
-import React, { ChangeEvent, FC, useContext } from "react";
+import React, { ChangeEvent, FC, useContext, useState } from "react";
 import { Box, Stack } from "@mui/material";
 import { StyledNativeSelect } from "../../../../Select.tsx";
-import { ReturnElements, StoreFunction } from "../../../QueryViewUtils.tsx";
+import {
+  formatXml,
+  getParserError,
+  ReturnElements,
+  StoreFunction
+} from "../../../QueryViewUtils.tsx";
 import TemplatePicker from "./TemplatePicker";
 import {
   QueryActionType,
@@ -11,23 +16,80 @@ import styled, { css } from "styled-components";
 import { TextField } from "@equinor/eds-core-react";
 import { Colors } from "../../../../../styles/Colors.tsx";
 import { useOperationState } from "../../../../../hooks/useOperationState.tsx";
+import { Button } from "../../../../StyledComponents/Button.tsx";
+import Icon from "../../../../../styles/Icons.tsx";
+import { DispatchOperation } from "../../../../../contexts/operationStateReducer.tsx";
+import OperationType from "../../../../../contexts/operationType.ts";
+import QueryService from "../../../../../services/queryService.ts";
+import ConfirmModal from "../../../../Modals/ConfirmModal.tsx";
 
-const QueryOptions: FC = () => {
+type QueryOptionsProps = {
+  onQueryChange: (newValue: string) => void;
+};
+
+const QueryOptions: FC<QueryOptionsProps> = ({ onQueryChange }) => {
   const {
     dispatchQuery,
     queryState: { queries, tabIndex }
   } = useContext(QueryContext);
   const {
-    operationState: { colors }
+    operationState: { colors },
+    dispatchOperation
   } = useOperationState();
 
-  const { storeFunction, returnElements, optionsIn } = queries[tabIndex];
+  const [isLoading, setIsLoading] = useState(false);
+
+  const { storeFunction, returnElements, optionsIn, query } = queries[tabIndex];
 
   const onFunctionChange = (event: ChangeEvent<HTMLSelectElement>) => {
     dispatchQuery({
       type: QueryActionType.SetStoreFunction,
       storeFunction: event.target.value as StoreFunction
     });
+  };
+
+  const validateAndFormatQuery = (): boolean => {
+    const formattedQuery = formatXml(query);
+    const parserError = getParserError(formattedQuery);
+    if (parserError) {
+      dispatchQuery({ type: QueryActionType.SetResult, result: parserError });
+    } else if (formattedQuery !== query) {
+      onQueryChange(formattedQuery);
+    }
+    return !parserError;
+  };
+
+  const sendQuery = () => {
+    const getResult = async (dispatchOperation?: DispatchOperation | null) => {
+      dispatchOperation?.({ type: OperationType.HideModal });
+      setIsLoading(true);
+      const requestReturnElements =
+        storeFunction === StoreFunction.GetFromStore &&
+        returnElements !== ReturnElements.None
+          ? returnElements
+          : undefined;
+      let response = await QueryService.postQuery(
+        query,
+        storeFunction,
+        requestReturnElements,
+        optionsIn?.trim()
+      );
+      if (response.startsWith("<")) {
+        response = formatXml(response);
+      }
+      dispatchQuery({ type: QueryActionType.SetResult, result: response });
+      setIsLoading(false);
+    };
+    const isValid = validateAndFormatQuery();
+    if (!isValid) return;
+    if (storeFunction === StoreFunction.DeleteFromStore) {
+      displayConfirmation(
+        () => getResult(dispatchOperation),
+        dispatchOperation
+      );
+    } else {
+      getResult();
+    }
   };
 
   const onReturnElementsChange = (event: ChangeEvent<HTMLSelectElement>) => {
@@ -69,12 +131,21 @@ const QueryOptions: FC = () => {
             })}
           </StyledNativeSelect>
         </Box>
-        <Box flexBasis="100%" textAlign="right">
+        <Stack
+          flexBasis="100%"
+          justifyContent="flex-end"
+          gap="0.5rem"
+          direction="row"
+        >
           <TemplatePicker
             dispatchQuery={dispatchQuery}
             returnElements={returnElements}
           />
-        </Box>
+          <Button onClick={sendQuery} disabled={isLoading}>
+            <Icon name="play" />
+            Execute
+          </Button>
+        </Stack>
       </Stack>
 
       {storeFunction === StoreFunction.GetFromStore && (
@@ -116,6 +187,26 @@ const QueryOptions: FC = () => {
       )}
     </Box>
   );
+};
+
+const displayConfirmation = (
+  onConfirm: () => void,
+  dispatchOperation: DispatchOperation
+) => {
+  const confirmation = (
+    <ConfirmModal
+      heading={"Delete object?"}
+      content={<span>Are you sure you want to delete this object?</span>}
+      onConfirm={onConfirm}
+      confirmColor={"danger"}
+      confirmText={"Delete"}
+      switchButtonPlaces={true}
+    />
+  );
+  dispatchOperation({
+    type: OperationType.DisplayModal,
+    payload: confirmation
+  });
 };
 
 const StyledTextField = styled(TextField)<{ colors: Colors }>`
