+++ Src/WitsmlExplorer.Frontend/styles/Icons.tsx
@@ -48,6 +48,7 @@ import {
   save,
   search,
   settings,
+  style,
   sync,
   text_field as textField,
   trending_up as isActive,
@@ -107,6 +108,7 @@ const icons = {
   save,
   search,
   settings,
+  style,
   sync,
   textField,
   tune,
